reorgonize code

+++ SConstruct
@@ -1,10 +1,78 @@
+#!/usr/bin/env python
+
 import os
+import platform
+import re
 import shutil
+import subprocess
 import sys
-from helpers import architectures, architecture_aliases, is_vanilla_clang, using_clang
 
 EnsureSConsVersion(3, 0, 0)
 EnsurePythonVersion(3, 6)
+
+architectures = ["x86_32", "x86_64", "arm32", "arm64"]
+architecture_aliases = {
+    "x86": "x86_32",
+    "x64": "x86_64",
+    "amd64": "x86_64",
+}
+
+
+def detect_arch():
+    host_machine = platform.machine().lower()
+    if host_machine in architectures:
+        return host_machine
+    elif host_machine in architecture_aliases.keys():
+        return architecture_aliases[host_machine]
+    elif "86" in host_machine:
+        return "x86_32"
+    else:
+        print("Unsupported CPU architecture: " + host_machine)
+        print("Falling back to x86_64.")
+        return "x86_64"
+
+
+def is_vanilla_clang(env):
+    if not using_clang(env):
+        return False
+    try:
+        version = subprocess.check_output(
+            [env.subst(env["CXX"]), "--version"]).strip().decode("utf-8")
+    except (subprocess.CalledProcessError, OSError):
+        print("Couldn't parse CXX environment variable to infer compiler version.")
+        return False
+    return not version.startswith("Apple")
+
+
+def get_compiler_version(env):
+    if not env.msvc:
+        try:
+            version = subprocess.check_output(
+                [env.subst(env["CXX"]), "--version"]).strip().decode("utf-8")
+        except (subprocess.CalledProcessError, OSError):
+            print("Couldn't parse CXX environment variable to infer compiler version.")
+            return None
+    else:  # TODO: Implement for MSVC
+        return None
+    match = re.search(
+        "(?:(?<=version )|(?<=\) )|(?<=^))"
+        "(?P<major>\d+)"
+        "(?:\.(?P<minor>\d*))?"
+        "(?:\.(?P<patch>\d*))?"
+        "(?:-(?P<metadata1>[0-9a-zA-Z-]*))?"
+        "(?:\+(?P<metadata2>[0-9a-zA-Z-]*))?"
+        "(?: (?P<date>[0-9]{8}|[0-9]{6})(?![0-9a-zA-Z]))?",
+        version,
+    )
+    if match is not None:
+        return match.groupdict()
+    else:
+        return None
+
+
+def using_clang(env):
+    return "clang" in os.path.basename(env["CC"])
+
 
 customs = ["custom.py"]
 opts = Variables(customs, ARGUMENTS)
@@ -12,7 +80,7 @@ opts.Add(BoolVariable("lto", "Link-time optimization", False))
 opts.Add(EnumVariable("target", "Target for build", "native_sdl",
                       ("linux_sdl", "windows_sdl", "native_sdl",
                        "linux_term", "windows_term", "native_term",
-                       "web", "tools", "tests")))
+                       "web", "tools", "prepare", "tests")))
 opts.Add(EnumVariable("arch", "CPU architecture", "auto",
          ["auto"] + architectures, architecture_aliases))
 opts.Add(
@@ -25,7 +93,11 @@ opts.Add(BoolVariable("clang", "Use clang", False))
 
 common_flags = ["-pipe", "-Wall", "-Wextra", "-pedantic"]
 env = Environment(CCFLAGS=common_flags, LINKFLAGS=common_flags,
-                  CPPPATH=['src', 'build', 'thirdparty', 'include'])
+                  CPPPATH=['src', 'build', 'thirdparty', 'include'],
+                  COMPILATIONDB_USE_ABSPATH=True
+                  )
+env.Tool('compilation_db')
+env.CompilationDatabase()
 opts.Update(env)
 env.msvc = False
 
@@ -98,50 +170,35 @@ if env["target"] == "web":
 env_tool = env.Clone()
 env_app = env.Clone()
 
-if not os.path.exists("build"):
-    os.mkdir("build")
-if not os.path.exists("build/assets"):
-    os.mkdir("build/assets")
+if not env["target"] == "prepare":
+    def get_source(path):
+        for file in os.listdir(path):
+            if os.path.isfile(os.path.join(path, file)):
+                if file.endswith('.c'):
+                    yield file
+    lib_src = []
+    list_dir = ["src", "src/base", "src/math"]
 
-os.system("./build/tools/png2c -n tsodinPog -o ./build/assets/tsodinPog.c ./assets/tsodinPog.png")
-os.system(
-    "./build/tools/png2c -n Lena_112 -o ./build/assets/Lena_112.c ./assets/Lena_112.png")
-os.system("./build/tools/png2c -n tsodinCup -o ./build/assets/tsodinCup.c ./assets/tsodinCup.png")
-os.system(
-    "./build/tools/png2c -n oldstone -o ./build/assets/oldstone.c ./assets/oldstone.png")
-os.system("./build/tools/png2c -n lavastone -o ./build/assets/lavastone.c ./assets/lavastone.png")
-os.system("./build/tools/obj2c -o ./build/assets/tsodinCupLowPoly.c ./assets/tsodinCupLowPoly.obj")
-os.system(
-    "./build/tools/obj2c -s 0.40 -o ./build/assets/utahTeapot.c ./assets/utahTeapot.obj")
-os.system("./build/tools/font2c -o ./build/assets/testFont.c -n test_font ./html/fonts/LibreBaskerville-Regular.ttf")
+    for dir in list_dir:
+        for file in get_source(dir):
+            lib_src.append(os.path.join(dir, file))
 
-
-def get_source(path):
-    for file in os.listdir(path):
-        if os.path.isfile(os.path.join(path, file)):
-            if file.endswith('.c'):
-                yield file
-
-
-lib_src = []
-list_dir = ["src", "src/base", "src/math"]
-
-for dir in list_dir:
-    for file in get_source(dir):
-        lib_src.append(os.path.join(dir, file))
-
-if not env["target"] == "web":
-    if os.path.exists("build/libs/libolive.so"):
-        os.remove("build/libs/libolive.so")
-    env_app.StaticLibrary(target='build/libs/olive', source=lib_src)
-else:
-    if os.path.exists("build/libs/libolive.a"):
-        os.remove("build/libs/libolive.a")
-    os.environ["EMSCRIPTEN_ROOT"] = os.path.dirname(WhereIs("emcc"))
-    env_app.Tool("emscripten", toolpath=[os.path.dirname(
-        WhereIs("emcc")) + "/tools/scons/site_scons/site_tools/emscripten"])
-    env_app.SharedLibrary(target='build/libs/olive',
-                          source=lib_src, LIBPATH='build/libs')
+    if not env["target"] == "web":
+        if os.path.exists("build/libs/libolive.so"):
+            os.remove("build/libs/libolive.so")
+        if env["target"] == "native_sdl" or "native_term":
+            env_app.StaticLibrary(
+                target='build/libs/olive', source=lib_src, CCFLAGS='$CCFLAGS -march=native')
+        else:
+            env_app.StaticLibrary(target='build/libs/olive', source=lib_src)
+    else:
+        if os.path.exists("build/libs/libolive.a"):
+            os.remove("build/libs/libolive.a")
+        os.environ["EMSCRIPTEN_ROOT"] = os.path.dirname(WhereIs("emcc"))
+        env_app.Tool("emscripten", toolpath=[os.path.dirname(
+            WhereIs("emcc")) + "/tools/scons/site_scons/site_tools/emscripten"])
+        env_app.SharedLibrary(target='build/libs/olive',
+                              source=lib_src, LIBPATH='build/libs')
 
 if env["target"] == "tests":
     env_tool.Program(target='build/test', source='test.c',
@@ -150,15 +207,56 @@ if env["target"] == "tests":
 elif env["target"] == "tools":
     env_tool.Program(target='build/tools/png2c',
                      source='tools/png2c.c', LIBS=['m'])
-    env_tool.Program(target='build/tools/obj2c',
-                     source='tools/obj2c.c', LIBS=['m'])
+    env_tool.Program(target='build/tools/obj2c', LIBPATH='build/libs',
+                     source='tools/obj2c.c', LIBS=['m', 'olive'])
     env_tool.Program(target='build/tools/font2c', source='tools/font2c.c', CCFLAGS='$CCFLAGS -I/usr/include/freetype2 -I/usr/include/libpng16 -L/usr/local/lib',
                      LINKFLAGS='$LINKFLAGS -I/usr/include/freetype2 -I/usr/include/libpng16 -L/usr/local/lib', LIBS=['m', 'freetype'])
+
+elif env["target"] == "prepare":
+    if not os.path.exists("build"):
+        os.mkdir("build")
+    if not os.path.exists("build/assets"):
+        os.mkdir("build/assets")
+
+    if os.name == 'nt':
+        os.system(
+            "./build/tools/png2c.exe -n tsodinPog -o ./build/assets/tsodinPog.h ./assets/tsodinPog.png")
+        os.system(
+            "./build/tools/png2c.exe -n Lena_112 -o ./build/assets/Lena_112.h ./assets/Lena_112.png")
+        os.system(
+            "./build/tools/png2c.exe -n tsodinCup -o ./build/assets/tsodinCup.h ./assets/tsodinCup.png")
+        os.system(
+            "./build/tools/png2c.exe -n oldstone -o ./build/assets/oldstone.h ./assets/oldstone.png")
+        os.system(
+            "./build/tools/png2c.exe -n lavastone -o ./build/assets/lavastone.h ./assets/lavastone.png")
+        os.system(
+            "./build/tools/obj2c.exe -o ./build/assets/tsodinCupLowPoly.h ./assets/tsodinCupLowPoly.obj")
+        os.system(
+            "./build/tools/obj2c.exe -s 0.40 -o ./build/assets/utahTeapot.h ./assets/utahTeapot.obj")
+        os.system(
+            "./build/tools/font2c -o ./build/assets/testFont.h -n test_font ./html/fonts/LibreBaskerville-Regular.ttf")
+    else:
+        os.system(
+            "./build/tools/png2c -n tsodinPog -o ./build/assets/tsodinPog.h ./assets/tsodinPog.png")
+        os.system(
+            "./build/tools/png2c -n Lena_112 -o ./build/assets/Lena_112.h ./assets/Lena_112.png")
+        os.system(
+            "./build/tools/png2c -n tsodinCup -o ./build/assets/tsodinCup.h ./assets/tsodinCup.png")
+        os.system(
+            "./build/tools/png2c -n oldstone -o ./build/assets/oldstone.h ./assets/oldstone.png")
+        os.system(
+            "./build/tools/png2c -n lavastone -o ./build/assets/lavastone.h ./assets/lavastone.png")
+        os.system(
+            "./build/tools/obj2c -o ./build/assets/tsodinCupLowPoly.h ./assets/tsodinCupLowPoly.obj")
+        os.system(
+            "./build/tools/obj2c -s 0.40 -o ./build/assets/utahTeapot.h ./assets/utahTeapot.obj")
+        os.system(
+            "./build/tools/font2c -o ./build/assets/testFont.h -n test_font ./html/fonts/LibreBaskerville-Regular.ttf")
 
 elif env["target"] == "native_sdl":
     for file in get_source("demos"):
         env_app.Program(target='build/demos/' + file.replace('.c', '') + '.sdl', source=(os.path.join("demos", file)),
-                        LIBPATH='build/libs', LIBS=['m', 'SDL2', 'olive'], CCFLAGS='$CCFLAGS -fno-builtin -DVC_PLATFORM=VC_SDL_PLATFORM')
+                        LIBPATH='build/libs', LIBS=['m', 'SDL2', 'olive'], CCFLAGS='$CCFLAGS -fno-builtin -DVC_PLATFORM=VC_SDL_PLATFORM -march=native')
 
 elif env["target"] == "linux_sdl":
     host_is_64_bit = sys.maxsize > 2**32
@@ -182,12 +280,12 @@ elif env["target"] == "windows_sdl":
 
     for file in get_source("demos"):
         env_app.Program(target='build/demos/' + file.replace('.c', '') + '.sdl', source=(os.path.join("demos", file)),
-                        LIBPATH='build/libs', LIBS=['m', 'SDL2', 'olive'], CCFLAGS='$CCFLAGS -fno-builtin -DVC_PLATFORM=VC_SDL_PLATFORM')
+                        LIBPATH='build/libs', LIBS=['m', 'SDL2', 'olive'], CCFLAGS='$CCFLAGS /fno-builtin /DVC_PLATFORM=VC_SDL_PLATFORM')
 
 elif env["target"] == "native_term":
     for file in get_source("demos"):
         env_app.Program(target='build/demos/' + file.replace('.c', '') + '.term', source=(os.path.join("demos", file)),
-                        LIBPATH='build/libs', LIBS=['m', 'olive'], CCFLAGS='$CCFLAGS -fno-builtin -DVC_PLATFORM=VC_TERM_PLATFORM -D_XOPEN_SOURCE=600')
+                        LIBPATH='build/libs', LIBS=['m', 'olive'], CCFLAGS='$CCFLAGS -fno-builtin -DVC_PLATFORM=VC_TERM_PLATFORM -D_XOPEN_SOURCE=600 -march=native')
 
 elif env["target"] == "linux_term":
     host_is_64_bit = sys.maxsize > 2**32
@@ -211,7 +309,7 @@ elif env["target"] == "windows_term":
 
     for file in get_source("demos"):
         env_app.Program(target='build/demos/' + file.replace('.c', '') + '.term', source=(os.path.join("demos", file)),
-                        LIBPATH='build/libs', LIBS=['m', 'olive'], CCFLAGS='$CCFLAGS -fno-builtin -DVC_PLATFORM=VC_TERM_PLATFORM -D_XOPEN_SOURCE=600')
+                        LIBPATH='build/libs', LIBS=['m', 'olive'], CCFLAGS='$CCFLAGS /fno-builtin /DVC_PLATFORM=VC_TERM_PLATFORM /D_XOPEN_SOURCE=600')
 
 elif env["target"] == "web":
     os.environ["EMSCRIPTEN_ROOT"] = os.path.dirname(WhereIs("emcc"))
